move mapping function to io/utils for io usage
